+++ lib/html.dart
@@ -7,9 +7,8 @@ library rikulo_html;
 import "dart:html";
 import "dart:math";
 
-import "view.dart";
 import "util.dart";
 
 part "src/html/Browser.dart";
 part "src/html/CSS.dart";
-part "src/html/DOMQuery.dart";
+part "src/html/DOMAgent.dart";