+++ lib/src/event/PopupEvent.dart
@@ -44,6 +44,6 @@ class PopupEvent extends ViewEvent {
       srcNode = source;
       popNode = popup is View ? popup.node: popup;
     }
-    return !new DOMQuery(srcNode).isDescendantOf(popNode);
+    return !new DOMAgent(srcNode).isDescendantOf(popNode);
   }
 }